+++ firmware/controllers/lua/lua.mk
@@ -9,7 +9,7 @@ ALLCPPSRC += $(LUA_DIR)/lua.cpp \
 			 $(LUA_DIR)/output_lookup.cpp \
 			 $(LUA_DIR)/value_lookup.cpp \
 			 $(LUA_DIR)/lua_can_rx.cpp \
-			 $(LUA_DIR)/lua_getchannel_hashes.cpp \
+			 $(LUA_DIR)/lua_getchannel.cpp \
 
 ALLINC += $(LUA_DIR) $(LUA_DIR)/luaaa $(LUA_EXT)
 ALLCSRC += \
