+++ firmware/controllers/lua/lua_getchannel.cpp
@@ -1,12 +1,8 @@
-#include "lua_getchannel_hashes.h"
+#include "lua_getchannel.h"
 #include "fuel_math.h"
 
-float getChannelByName(const char *name)
-{
-	auto hash = djb2lowerCase(name);
-
-	switch (hash)
-	{
+float getChannelByName(const char* name) {
+	switch (djb2lowerCase(name)) {
 	case djb2lowerCase("FuelFlow"):
 		return engine->module<TripOdometer>()->getConsumptionGramPerSecond();
 	case djb2lowerCase("AFR"):
@@ -18,6 +14,4 @@ float getChannelByName(const char *name)
 	default:
 		return 0.0f;
 	}
-
-	return 0.0f;
 }
