+++ firmware/controllers/lua/lua_hooks.cpp
@@ -10,7 +10,7 @@
 #include "lua_airmass.h"
 #include "value_lookup.h"
 #include "can_filter.h"
-#include "lua_getchannel_hashes.h"
+#include "lua_getchannel.h"
 #if EFI_CAN_SUPPORT || EFI_UNIT_TEST
 #include "can_msg_tx.h"
 #endif // EFI_CAN_SUPPORT
